refactor(forge): split Forgejo label methods into forgejo_labels.go (#1025)
